Worked on the feedback given. Cleaned the code a bit.
